+++ packages/react-auth/src/AuthProvider/index.tsx
@@ -2,6 +2,7 @@
 
 import {
   FC,
+  type JSX,
   createContext,
   useCallback,
   useContext,