+++ packages/react-mui-audit-log-viewer/rollup.config.mjs
@@ -1,16 +1,15 @@
-// /rollup.config.js
-import resolve from "@rollup/plugin-node-resolve";
-import external from "rollup-plugin-peer-deps-external";
-import terser from "@rollup/plugin-terser";
-import postcss from "rollup-plugin-postcss";
-import typescript from "@rollup/plugin-typescript";
 import commonjs from "@rollup/plugin-commonjs";
-import replace from "@rollup/plugin-replace";
-import dts from "rollup-plugin-dts";
 import json from "@rollup/plugin-json";
+import resolve from "@rollup/plugin-node-resolve";
+import replace from "@rollup/plugin-replace";
+import terser from "@rollup/plugin-terser";
+import typescript from "@rollup/plugin-typescript";
+import dts from "rollup-plugin-dts";
+import external from "rollup-plugin-peer-deps-external";
 import nodePolyfills from "rollup-plugin-polyfill-node";
+import postcss from "rollup-plugin-postcss";
 
-import pkg from "./package.json";
+import pkg from "./package.json" assert { type: "json" };
 
 export default [
   {
@@ -20,6 +19,7 @@ export default [
         dir: "dist/cjs",
         format: "cjs",
         sourcemap: true,
+        interop: "auto",
         name: "react-lib",
       },
       {