+++ packages/react-mui-branding/rollup.config.mjs
@@ -1,14 +1,14 @@
-import resolve from "@rollup/plugin-node-resolve";
-import external from "rollup-plugin-peer-deps-external";
-import terser from "@rollup/plugin-terser";
-import postcss from "rollup-plugin-postcss";
-import typescript from "@rollup/plugin-typescript";
 import commonjs from "@rollup/plugin-commonjs";
-import nodePolyfills from "rollup-plugin-polyfill-node";
-import dts from "rollup-plugin-dts";
 import json from "@rollup/plugin-json";
+import resolve from "@rollup/plugin-node-resolve";
+import terser from "@rollup/plugin-terser";
+import typescript from "@rollup/plugin-typescript";
+import dts from "rollup-plugin-dts";
+import external from "rollup-plugin-peer-deps-external";
+import nodePolyfills from "rollup-plugin-polyfill-node";
+import postcss from "rollup-plugin-postcss";
 
-import pkg from "./package.json";
+import pkg from "./package.json" assert { type: "json" };
 
 export default [
   {
@@ -39,7 +39,7 @@ export default [
     external: Object.keys(pkg.peerDependencies || {}),
   },
   {
-    input: "dist/esm/types/index.d.ts",
+    input: "src/index.ts",
     output: [{ file: "dist/types/index.d.ts", format: "esm" }],
     external: [/\.css$/],
     plugins: [dts()],